requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
openai==1.51.0
requests==2.31.0
yfinance==0.2.28
pytest==7.4.4
pytest-asyncio==0.21.1
